This checking re-writes the function factorf in poly/facform.red, which could previously deliver incorrect results. To try to reduce future confusion it renames the old function to be interbal!-factord and makes the name factorf a synonum for the main entrypoint (fctrf). A modest number of other files that refer to factorf directly have then been changed to track this.
Also in int/dint.red the driver for definite integration has a new segment enabled by "on acn;". This is not a long-term change, but an experiment and it tries
using definite integration followed by simple substitution as an earlier
stage of attempting indefinite integration. It is possible that the reported
integration bugs that prompted this have been fixed by the facform change and
by other good changes that Eberhard and Rainer have or are checking in!