First note that this checkin only changes rlisp/proc.red and alg/reval.red. proc.red puts file-name and line number on the property list of procedure names as they are defined. The way it does this is such that the info will only tend to survive during the current run, so a saved Reduce image is not overloaded with this info. reval.red hooks onto the code where user-defined algebraic procedur4es are called (specifically when the 'opfn property of an operator is used to cause a procedure to be called). If you go    algebraic procedure foo x; x + 1;    tralg foo; and then use foo you may see some trace output! Sort of obviously untralg foo; would untrace it. Right now as well as an attemt at an algebraic display of args and results there is a Lisp prefix display - that is intended to be temporary while this feature is explored and tested. With luck (and I have not tested it al all!) if you have "on backtrace" then utter failures within algebraic procedures may lead to a bit of a backtrace. Since I have not tried that I do not know how satisfactory it will seem.
It may be that where I hooked this in is not the only place something like
this is "needed", or that the information I display is inadequate or ugly.
Well I know of several ways in which it is not as nice as I could imagine -
but before looking at this any more I want to have people check if my changes
break things and look for (simple please) ways it could be enhanced. Ideally
by looking at the implementation and proposing changes or updates. After all
what I have at present is not very many lines of code!

Well HMMM in fact right now this seems to allow the Reduce resrs to run OK
apart from specfn which seems to get stuck. I will still check this in this
evening and look into that issue tomorrow!