A couple of changes to the defint package:  I replaced explicit cons'ing of !*sq by a call to mk!*sq,  rewrote some inequality tests to use the more general simp!-sign  instead of explicitly trying to calculate both sides in rounded mode.
Further, I noticed that
realvalued a;
int(e^(-a*x),x,0,infinity);

did always return 1/a, irregardless of the sign of a.

Found that the convergence conditions for the MeijerG integral did calculate
angle of a comples number a incorrectly, namely as atan(impart(a)/repart(a))
which gives wrong values in left left halfplane and tries to divide by zero
for purely imaginary a.

Replaced this by the correct atan2(impart(a)/repart(a)).

After this change, the integral

 int(x^(-1)*sin(x)*cos(x/2),x,0,infinity);

suddenly fails. Again this seems to be a problem in testing the convergence   
conditions: after exchanging the factors sin(x) and cos(x/2) the integral
is calculated correctly.  Compare:

 defint2(1/x,sin(x),cos(x/2),x);

and

 defint2(1/x,cos(x/2),sin(x),x);

This needs further investigation.    

Added a couple of tracing statements for some of the convergence conditions.