In CSL there had been a garbage-collection related bug. I believe it just arose when there was a function with at least four arguments that called itself (it may have been that only tail-calls would hurt), and then that function was heavily enough used to end up compiled into C++, and finally a garbage collection got triggered precisely at the point of that self-call. The consequence could be an illegal memory reference (SIGSEGV) because a value that was needed was not kept safe in a place where the garbage collector knew to update it when it copied all the heap. I hope that this checkin fixes things, because the bug was horrible to track down, perhaps because very small changes in environment could lead to a garbage collection being triggered one step eariler or one step later than the place that led to failure.
I have also improved (I hope) "scripts/csl-sanity-check.sh" so it is less
sensitive about just what set of fonts you have installed on an X-server.
And fixed a couple of typos in comments. I am part way to making the Lisp to
C++ converter translate rplaca and rplacd inline but that does not yet work
and is not yet enabled.