Standard LISP on Emacs Lisp updates
The previous version did not process "FOR EACH ... DO" correctly due to mishandling of LAMBDA, which also prevented the REDUCE LAMBDA statement from working. Having spent the last couple of weeks experimenting, I have concluded that there is no feasible way to handle LAMBDA as an upper-case symbol when only the lower-case symbol lambda is defined in Emacs lisp. Instead, I now use the newnam and quotenewnam facilities together (both are necessary) in boot.el to convert LAMBDA in REDUCE source code to lambda when it is read, and this approach seems to work.

Added debugging information to macros that I missed before and added comments to explain why macros are required instead of aliases. (Aliases either don't work at all or prevent compilation of boot.el.) I now implement EVAL, APPLY and FUNCTION as aliases since I no longer try to process LAMBDA specially, which seems to work. However, QUOTE needs to be a macro rather than an alias!

Tidied up the recently added read-eval-print loop, but that still needs further work because output tends to end up in the *Messages* buffer and not where it should be!