Standard LISP on Emacs Lisp updates
Hacked COMPRESS to precede a leading colon in an identifier with !, so that Emacs Lisp does not regard it as a keyword and complain when compiled code uses functions with such names.

Added standard Emacs Lisp header comments and revised a few other comments.

Expanded the macro `with-current-buffer' in function OPEN because this adds very little extra complexity and makes the code more transparent.

Tweaked the UI code to echo minibuffer input and look slightly nicer, but I still can't solve the problem of output going astray!