Standard LISP on Emacs Lisp updates
Backported to Emacs 25 (which is the currently released version).

I have also located the cause of output going astray. I think it is an Emacs bug, which I have reported. A user-error in the minibuffer, such as trying to access a nonexistent history item, sets standard-output globally to t, thereby redirecting all print output to the *Messages* buffer! In the meantime, I have added a workaround to reset standard-output after each call of (read), which is probably sensible defensive programming anyway.