Standard LISP on Emacs Lisp updates
Arrange to define the CX..XR compiler macro at compile time so that calls to CX..XR functions in sl.el (used for Emacs 25 compatibility) compile correctly.

Only retain the !: at the start of an identifier if it is followed by a letter, so as not to mess up `:', `:=', `: ', etc.

Allow a global true value of debug-on-error to override the tr argument to ERRORSET to facilitate debugging.

Implement the LINELENGTH function, which is required for algebraic-mode output.

Implement READ, which is required for the YESP predicate. But I have only implemented upcasing for identifiers, which is all that is required to support YESP.

Echo minibuffer input from READ and READCH, which otherwise disappears. Prevent minibuffer errors resetting standard-output more elegantly. Correctly return $EOL$ when READCH reads interactive input via the minibuffer.

I can now process at least some of "alg.tst" correctly.