Standard LISP on Emacs Lisp updates
In COMPRESS: Retain a leading !: followed by a digit to prevent the Standard LISP identifier being an Elisp keyword. (Such an identifier appears in the arith package.) Emacs does not accept .E as in 123.E-2 so delete the ".". (This is the form that REDUCE seems to generate when it reads a float.)

FLUID and GLOBAL now call defvar to define Elisp free variables properly.