Standard LISP on Emacs Lisp updates
In COMPRESS: !\ preceding an identifier is removed and the identifier is downcased to allow direct use of Emacs Lisp functions.

Convert PUT to a macro. This generates better compiled code than defalias by referencing only put, not both put and PUT.

Suppress compiler warnings about use of undeclared free variables in DE, PUTD and the fasl support functions.

Tidy up GETD and support compiled macros better. However, this may need further work because the forms of macros are different in Emacs and Standard Lisp and I currently don't try to convert compiled code (which I'm not sure is even feasible).

Revise PUTD to try to support non-lambda (e.g. compiled) function forms.

Add some more functions that are not in the Standard LISP report but are assumed by REDUCE, namely RASSOC, EXPLODE2 and INT2ID. Also define LOAD-MODULE and EVLOAD in sl.el so that I can use them early in the boot process, and because it's easier to define them in Lisp than Rlisp, LOAD-MODULE is ESL-specific and EVLOAD uses LOAD-MODULE.

Modified FASLOUT/FASLEND to work with the REDUCE 3.8 boot code (but I think I will also modify the boot code, which is somewhat ESL-specific anyway, to support my model of generating fasl files better).