Standard LISP on Emacs Lisp updates
Prevent the use of ON DEFN from changing property lists by saving and restoring them.  Because ESL REDUCE essentially uses ON DEFN to generate fasl files, this means that compiling modules no longer mangles properties, which was messing up the build process for at least one module.

The issue is that DEFLIST, FLAG, FLUID, GLOBAL, REMFLAG, REMPROP and UNFLUID are flagged EVAL and so are evaluated (as well as being displayed) with ON DEFN, but PUT is not. This means that inputting some REDUCE files with ON DEFN can remove properties (so that the file can be read correctly) but not reinstate them, so that subsequently REDUCE does not work correctly.

Move the suppression of compiler warnings to the right places so as not to interfere with normal Emacs Lisp compilation.

Revert to using Emacs prettyprinted (rather than simply printed) Lisp output when generating fasl files, and keep it for debugging. (But this needs further work because, at least by default, the Emacs Lisp prettyprinter doesn't format Standard Lisp anywhere near as well as the REDUCE prettyprinter does.)