Standard LISP on Emacs Lisp updates
Include a Standard LISP prettyprinter, primarily to produce readable Lisp files for debugging purposes when generating fasl files.  The new file "eslpretty.el" is the file "packages/rprint/pretty.red" converted to Lisp syntax (using ESL REDUCE!), then modified slightly to be independent of REDUCE so that it is available from the start of the REDUCE boot process. The compiled version of this file is loaded using `require' when generating fasl files or when the switch DEFN is turned on explicitly.

Disable compilation of macros defined using DM, which is generating bad code, in particular for REPEAT and WHILE. Fix some minor errors in FLUID and GLOBAL and one major error. FLUID is usually called `statically' with a quoted list of arguments but it is also called dynamically (e.g. in SWITCH) with a computed list of arguments, in which case the previous version did not work. I now handle these two cases separately, which seems to work. But FLUID and GLOBAL could probably both be handled more elegantly.

Handle the naming of IO buffers more elegantly to try to avoid a repeat of the error I caused by renaming the prefixes without allowing for the change in prefix length. Some further minor tidying up.

I believe these files build (at least most of) REDUCE.