Standard LISP on Emacs Lisp updates
DM and PUTD now save as a property and GETD restores source versions of Standard Lisp macro definitions, since I'm not convinced that my attempt to convert macros between Standard and Emacs Lisp forms was working correctly, especially for compiled macros. This approach should be safer, at the expense of using more memory.

My current build of ESL REDUCE works as well as any previous build: it will do some calculations but not others! But at least I now have a prettyprinted Lisp source code version of each fasl file, which should make debugging easier.

Add a check (not yet tested) that the Lisp files output as part of generating fasl files are readable and not empty, which is still the case for a few files later in the build process.