GNU Emacs REDUCE IDE
In reduce-run.el: Allow the reduce-run-commands option to be customized before reduce-run has been fully loaded.  When sending REDUCE code from a file to a REDUCE process, don't switch to the process window unless this is explicitly requested (by an interactive argument). This was supposed to be the behaviour, but I think a fairly recent round of changes must have broken it. This change tidies up the code quite a lot.