REDUCE on Emacs Lisp
Revise QUOTIENT applied to (possibly big) integers.  The way Calc works by default with negative arguments is not appropriate.

Add Edebug support to the PLUS and TIMES macros.

Use the Elisp version of boundp instead of the REDUCE definition.  The latter intentionally triggers an error with an unbound argument, which makes debugging awkward.

The "poly.tst" file now appears to run correctly if I replace the let rules with procedures, i.e. the decompose and gcd tests themselves now seem to succeed but there is a problem with the way this test file uses let.