Building Emacs REDUCE
This check-in tidies up a few loose ends.

In "esl.el": Correct the handling of identifiers beginning with : so as not to mangle the REDUCE prompt.  Improve READCH so that it handles end-of-line better.  Batch-mode input now works correctly and there are no spurious blank lines in echoed input.  This change lets me write "remake-one-package.sh" as I want to.  I have also added an attempt at an automatic check for errors to the end of "remake-all-packages.sh".