Emacs Standard Lisp
GETD now accepts any argument type and returns nil if it is not an identifier, as per the Standard Lisp Report.

PROG now deletes all null sexprs in its body.  These can be generated by the REDUCE parser and upset my implementation using the Emacs Common Lisp emulation library.  (I already knew that a final null sexp could be generated and deleted that, but a penultimate null sexp is generated by a while loop with an empty body, so now I delete all null sexps.)

These two long-standing errors were shown up by the REDUCE randpoly test, which now runs correctly except for a possible display bug that only seems to appear in batch mode!  This might be related to another general display bug where some output is move too far to the right, which is clearest in the misalignment of matrix delimiters.