ESL
WRS no longer regards the filehandle (t . OUTPUT) as an error.  It can occur when output goes to stdout in batch mode.  This was picked up by the gentran and fide test files.

These two test files also show errors in the handling of the identifier !T in FORTRAN output, and I think the fide test output is also more significantly wrong because identifiers containing t are not sorting correctly.  These are consequence of preserving t as lower-case in what is generally an upper-case Lisp.  Clearly I need to be more careful how I do this!