Emacs REDUCE
Comment out flagging of load (and reload) in "eslrend.red", as in "pslrend.red", so that normally formload gets called (and treats load as load-package).  This fixes the geoprover test.

Now that LOAD does not throw an error for a missing file, no need to turn off debug-on-error when building plot.