Minor correction to rlisp/tok.red
The first return statement in the definition of read!-comment at the end of the file was misread as a variable in the initial bootstrapping process because the code to read it correctly is in rlisp/block.red, which is read AFTER tok.red.  I have therefore added an explicit argument of nil, which seems to fix the problem.