REDUCE on Common Lisp
I now support REDUCE on CLISP to the same extent that I support it on
SBCL.  This involve using a combination of more robust code and
customisation.

CLISP is significantly slower than SBCL!

Since CLISP is only readily available for Windows under Cygwin, I
support both the current Cygwin version of gnuplot and the (rather
old) native Windows version that is included in the Windows
distribution of REDUCE.  I run the Cygwin version if it is available
and fall back to the native Windows version otherwise.  Currently, I
only support dumb tty output with Cygwin gnuplot but supporting
graphical output should be trivial, and in the longer term I should be
able to experiment with REDUCE using the current version of gnuplot.