Minor correction to pgauss test file
Running the file pgauss/pgauss.tst produces syntax errors and general nonsense for the last third.  The reason appears to be that the code is in symbolic mode but is being processed in algebraic mode.  I have added a symbolic declaration where I think it should be and now the code produces what look like reasonable results.