Lower-case REDUCE on Common Lisp
REDUCE on Common Lisp is now lower-case and case-insensitive, as is REDUCE on PSL/CSL.  Currently this only applies to REDUCE on CLISP, which provides a "modern" option that is case-sensitive and case-inverting.

Making REDUCE lower-case fixes most of the remaining problems and removes the need for several hacks.  I have also generalized compress and explode to handle lists and compress to ignore garbage after an object, which is assumed by a couple of packages.  REDUCE on CLISP now runs nearly all the standard test files; see the README file for details.  The biggest problem is that it is very slow.  One of the things I need to think about is porting the lower-case support to SBCL, which seems to be much faster.