REDUCE on Common Lisp
Provide lower-case REDUCE on SBCL that works slightly better than the
previous CLISP version and is a lot faster.  It uses a portable
CS-COMMON-LISP package to provide most of the relevant case-inversion
functionality of the CS-COMMON-LISP package available in CLISP.
Essentially, intern inverts the case of a symbol name and symbol-name
inverts it back again.  Going through this additional layer makes SBCL
REDUCE a little slower than the previous version.  The case-inversion
code now needs to be absorbed into the STANDARD LISP package in a
portable, elegant and efficient way.

Beware that I have only tested this version on SBCL and it may not run
correctly on CLISP.