REDUCE on Common Lisp now runs on Linux
This version builds and runs similarly on both MS Windows and Linux. I
have only tested it on Ubuntu but it ought to run with little or no
modification on any Unix-like platform that provides a bash shell.  It
should build within the current Subversion file tree with no
additional configuration.  REDUCE can be run via the command redsbcl
or redclisp from any directory and the current directory can be
changed while REDUCE is running using the command cd (or its alias
chdir).

SBCL REDUCE is currently around 3 times slower than CSL REDUCE and
CLISP REDUCE is around 15 times slower than CSL REDUCE.  This is with
compilation optimized for speed in both cases.

SBCL and CLISP REDUCE now use independent sub-directories of the
common-lisp directory.  The REDUCE image files now include start-up
code that launches REDUCE immediately; it no longer runs on top of a
Lisp REPL.

I have added type declarations for functions that accept or return
restricted types, but this has not improved the speed noticeably.

My next step will probably be to try to improve the error handling,
which is currently a mess.