REDUCE web site
Redirect only .htm (and not also .html) URLs to .php.  (This fixes a
bug in the previous check-in.)

Build the list of links to user-contributed package documents
dynamically from the index page for the HTML version of the REDUCE
User's Manual.  This should automatically stay up to date, whereas the
previous list of links to PDF documents is now somewhat out of date.
The code for this is all in a new template file for the documentation
page.

If the markup of the HTML REDUCE User's Manual changes significantly
then the JavaScript that builds the list will need updating!  (It
doesn't use the chapter or section numbering.)  Also, I have had to
rewrite o-umlaut, which otherwise is not rendered correctly.  I think
this is because the HTML REDUCE User's Manual uses an old character
encoding whereas the rest of the web site uses UTF-8, although the
code point is the same in both encodings.  This issue may need further
attention if any other non-ASCII characters appear in the titles of
user-contributed packages.