REDUCE web site: add convenience links
Convert the bee icon at the left of the header into a link to the
project page.  Add a button at the right of the header that links to
the new search page but hide it on the search page itself.  Add a
search icon to the Search item in the navigation menu.  Add a
statement at the top of the search page to explain what results the
search will show.