REDUCE web site for small displays
Make the background for REDUCE in the header responsive.  Use media
queries to replace the background image by a simple coloured
background for medium and smaller displays, and set the width of the
simple coloured background to be 60% of the viewport width (60vw).  (I
can't see any way to do this using bootstrap facilities so I'm using
CSS directly.)  This keeps the menu icon on the header line on a
display down to 318px wide.