Two unrelated changes in this revision:
misc/dfpart
Correct obvious typo in rule for dfp of a sum in dfpart.red.
Additions to dfpart.tst to illustrate the effect of this and
the resultant version of dfpart.rlg.

specfn/sfellipi.red
Use theta functions evaluated at zero rather than direct series summations
in the numerical evaluation of nome2mod, nome2!K etc as modular transformations
can be used to speed up convergence when |q| is near to one.