REDUCE web site: various minor updates
Use a smaller square marker to indicate which menu item is selected or
about to be selected.  This is more elegant and what I always
intended.  It is also closer to the styling of Tony's original
reduce-algebra.com site.

Add a facility to allow me to validate pages more easily.  Fix invalid
markup on several pages.  (It's amazing how forgiving web browsers
are!)

Update the text on the obtaining and windows pages to agree better
with current reality.