Catastrophic error in num_int when factor is on:
on factor;
u := 1/(sqrt(sin(x))+sqrt(cos(x)))^4;
load_package numeric;
num_int(u,x=(0.0 .. pi/2.0));

Solved by binding !*factor to nil in boundseval.

With on complex, the same input caused a "car of an atom" error.
This was a problem with the simpfn for complex-rounded domain elements,
which expected a domain element as argument but was called with the tag
already stripped.

Solved by flagging !:cr!: as 'full.