Thanks to Francis Wright for two very useful hints used in this revision.
decrep.red:  Changes to allow display of the periodic part of a decimal by
using overlining on platforms that support this: namely Run-Reduce-FX
and TeXmacs.

sfellip.red & sfellipi.red:
First attempt to bring the notaion used for the output of elliptic functions
into line with the NIST Digital Library of Mathematical Functions.
The Jacobi and Legendre incomplete integrals of the second kind are now
denoted by JacobiE and EllipticE respectively.

Also changes to printing of JacobiE using \mathcal{E} on platforms that
support this. Otherwise it is printed as !Epsilon or \epsilon depending on
whether FANCY is OFF or ON.

rataprx.tst & sfellip.tst: updated to reflect these changes together with
updated .rlg files.

Changes to the manual files rataprx.tex and specfn.tex to document the above.