The two recent issues with the defint package are now resolved.
In defintb.red, gamma was called with two parameters which fell foul of the new
argument checking as gamma is unary.  This was changed to m_gamma (the
unnormalised incomplete gamma function) by Rainer Schoepf in subversion 5574.

Similarly in definte.red, hankel2 was called with three parameters instead of
two.  As suspected, this was due to the overloading of Hankel2 in the defint
package to represent the Hankel transform as well as a call to the Bessel
function of the third kind, Hankel2.  This has now been resolved by renaming
hankel2 as hankelt in files defintd.red and definite.red and declaring hankelt
to be an operator.  The resulting file defint.rlg is unchanged from that in
subversion 5576.

dilog.red & sfbes.red:
The number!-of!-args properties of gamma and Hankel2 have been reset to their
correct values 1 and 2 respectively.