turtle.tex
In the HTML version, the tabbing environment is currently displayed as
an image, which is incorrectly attached to the preceding line, so I
have added a blank line to make it a separate paragraph.  This fixes
the problem and doesn't seem to affect the PDF version.  I have also
fixed some typographical errors.