Proof of concept for a MathView based web-interface. It shows that the input equation editor is functional. Cut and paste of 2-dimensional expressions work. Currently, only polynomials and sqrt's are supported as no other serializers are written yet. Unfortunately, to try this out a reduce web image has to be build with the following modification to csl.cpp
#elif defined PROCEDURAL_WASM_XX
// Set up Reduce
    PROC_prepare_for_top_level_loop();
    PROC_process_one_reduce_statement("off echo$");
    PROC_process_one_reduce_statement("algebraic$");
    PROC_process_one_reduce_statement("off nat$");
// Here I initialise state in the way that we do when we are going to run
// under TeXmacs.
//    PROC_process_one_reduce_statement(
//        "<< lisp (!*redeflg!* := nil)$ load_package tmprint$"
//        " on fancy$ off promptnumbers$ on redfront_mode >>$");
//    PROC_process_one_reduce_statement("int(x/ln(x^2-beta^2), x);");

Now, if you issue ./launch__MathLiveTests.sh you can type, for example, (a^2 -3)^2; into the input field of the browser. Then push the submit button. Both input and output will be displayed typeset and are editable. Selections of the displayed expressions can also be copied and pasted (with ctrl-c , ctrl-v on my browser).

The next step will be to try to use MathJSON as an exchange format. That would make serialization trival.