HTML manual fixes
Change the bookmark driver default to dvipdfm instead of dvips because
the latest version of the latter, namely bkm-dvips.def 2020-11-06
v1.29, interacts badly with dvipng causing a floating point exception
every time it is called by TeX4ht.  (I have reported this bug and it
will be fixed in the next release of dvipng, presumably version 1.18.)

Introduce a new command to include the image files in turtle.tex that
avoids rotating them when TeX4ht is in use.  The rotation caused a
GhostScript error and failed, but no rotation of the EPS images is
necessary anyway.