Web site
General somewhat-overdue update, including updating the Bootstrap,
jQuery and MathJax JavaScript libraries to their latest stable
released versions.  I have replaced downloading.php and windows.php
with building.php, which gives a quick general outline of building
REDUCE but defers to README.BUILDING in the repository for the
details.  It needs a little more work to make links to sections of the
HTML manual reliable and I would like to add a site map.