Manual revision
Revise the chapter on "Commands for Interactive Use": expand the
section on "Referencing Previous Results" to clarify exactly what WS
references; link to the section on "The Expression Workspace" in the
chapter on "Display and Structuring of Expressions"; use "you" instead
of "the user" or "he"; add keywords to section titles.  Replace $Z$
with \textit{Z} in the title of the subsection on the Z-transform
package to render it correctly on the documentation web page without
the need to run MathJax.