Manual: Z-transform package
Replace $Z$ with \textit{Z} in the title of the subsection on the
Z-transform package to render it correctly on the documentation web
page without the need to run MathJax.  But this time, make the change
in the right file, namely user.tex, to always use \textit{Z}, and
revert the previous change in ztrans.tex.