Web site: external linking to the manual
Fix the flash of underlying page by setting the initial body style to
"display:none" and resetting it to "display:inherit" only if the page
is being accessed without a query string.  This seems to work OK as a
prototype, but it really needs to be re-implemented in PHP and run on
the server rather than the client.