Web site
This is the second step of a revision of the web site to stop using
the Smarty template engine and just use "raw" PHP.  I have now
re-implemented and checked all the subsidiary pages, and deleted the
smarty template directories.  In particular, the tutorial pages should
be much more intelligible and maintainable since the content is now
straightforward HTML rather than nested PHP arrays.  I have kept the
HTML as simple as possible and used jQuery to add the more complex
markup that Bootstrap requires for the concertina animation.  I have
also revised one link on the index page.